+++ lib/fur/core/lists.rb
@@ -5,9 +5,9 @@ module Fur
       extend Runtime::RustLibrary
 
       rust('libfur') do
-        ff('head', [Runtime::List], Runtime::Integer)
-        ff('tail', [Runtime::List], Runtime::List)
-        ff('last', [Runtime::List], Runtime::Integer)
+        ff('head', [Runtime::List::Integer], Runtime::Integer)
+        ff('tail', [Runtime::List::Integer], Runtime::List)
+        ff('last', [Runtime::List::Integer], Runtime::Integer)
       end
 
       define_function(:head, [Runtime::Param.new(:a, :list)]) do |scope|
@@ -19,7 +19,7 @@ module Fur
       define_function(:tail, [Runtime::Param.new(:a, :list)]) do |scope|
         list = scope.get(:a)
         raise Runtime::EmptyList.new(:tail) unless list.any?
-        Runtime::List.from_ff(tail.call(list.to_ff))
+        Runtime::List::Integer.from_ff(tail.call(list.to_ff))
       end
 
       define_function(:last, [Runtime::Param.new(:a, :list)]) do |scope|
@@ -30,7 +30,7 @@ module Fur
 
       define_function(:map, [Runtime::Param.new(:a, :list), Runtime::Param.new(:b, :fn)]) do |scope|
         list, function = scope.get(:a).call(scope), scope.get(:b).call(scope)
-        Runtime::List.new(
+        Runtime::List::Integer.new(
           list.map do |member|
             Runtime::Exec.new(function, [member.call(scope)]).call(scope)
           end
